D:作業完成，新增資料並存進DB中

+++ MVCHomework/Models/ViewModels/AddBalanceViewModel.cs
@@ -25,7 +25,7 @@ namespace MVCHomework.Models.ViewModels
         /// 紀錄金額
         /// </summary>
         [DisplayName("紀錄金額")]
-        public decimal BalanceMoney { get; set; }
+        public int BalanceMoney { get; set; }
         /// <summary>
         /// 紀錄時間
         /// </summary>
